+++ core/src/main/java/be/orbinson/aem/dictionarytranslator/servlets/action/ReplicateDictionaryServlet.java
@@ -1,15 +1,15 @@
 package be.orbinson.aem.dictionarytranslator.servlets.action;
 
+import com.day.cq.replication.ReplicationActionType;
+import com.day.cq.replication.ReplicationException;
+import com.day.cq.replication.Replicator;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.SlingHttpServletResponse;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.servlets.SlingAllMethodsServlet;
-import org.apache.sling.distribution.DistributionRequest;
-import org.apache.sling.distribution.DistributionRequestType;
 import org.apache.sling.distribution.Distributor;
-import org.apache.sling.distribution.SimpleDistributionRequest;
 import org.apache.sling.servlets.annotations.SlingServletResourceTypes;
 import org.jetbrains.annotations.NotNull;
 import org.osgi.service.component.annotations.Component;
@@ -17,6 +17,7 @@ import org.osgi.service.component.annotations.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.jcr.Session;
 import javax.servlet.Servlet;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -34,6 +35,11 @@ public class ReplicateDictionaryServlet extends SlingAllMethodsServlet {
     @Reference
     private transient Distributor distributor;
 
+    @Reference
+    private transient Replicator replicator;
+
+    private ResourceResolver resourceResolver;
+
     @Override
     protected void doPost(SlingHttpServletRequest request, @NotNull SlingHttpServletResponse response) throws IOException {
         String path = request.getParameter("path");
@@ -42,23 +48,35 @@ public class ReplicateDictionaryServlet extends SlingAllMethodsServlet {
             LOG.warn("Invalid parameters to replicate dictionary");
             response.sendError(HttpServletResponse.SC_BAD_REQUEST);
         } else {
-            final ResourceResolver resourceResolver = request.getResourceResolver();
+            resourceResolver = request.getResourceResolver();
             Resource resource = resourceResolver.getResource(path);
 
             if (resource != null) {
-                DistributionRequest distributionRequest = new SimpleDistributionRequest(DistributionRequestType.ADD, true, path);
-                distributor.distribute("publish", resourceResolver, distributionRequest);
+                deepReplicate(resource);
 
                 if (LOG.isDebugEnabled()) {
-                    // javasecurity:S5145
                     LOG.debug("Replicated dictionary, 'path={}'", path.replaceAll("[\n\r]", "_"));
                 }
             } else {
                 if (LOG.isWarnEnabled()) {
-                    // javasecurity:S5145
                     LOG.warn("Unable to get resource for path '{}", path.replaceAll("[\n\r]", "_"));
                 }
                 response.sendError(HttpServletResponse.SC_BAD_REQUEST);
+            }
+        }
+    }
+
+    private void deepReplicate(Resource parentResource) {
+        String path = parentResource.getPath();
+        try {
+            replicator.replicate(resourceResolver.adaptTo(Session.class), ReplicationActionType.ACTIVATE, path);
+        } catch (ReplicationException e) {
+            LOG.error("ReplicationException occurred when trying to replicate dictionary in ReplicateDictionaryServlet");
+        }
+
+        if (parentResource.hasChildren()) {
+            for (Resource childResource : parentResource.getChildren()) {
+                deepReplicate(childResource);
             }
         }
     }