+++ core/src/main/java/be/orbinson/aem/dictionarytranslator/servlets/action/UpdateLabelServlet.java
@@ -5,7 +5,11 @@ import com.day.cq.commons.jcr.JcrUtil;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.SlingHttpServletResponse;
-import org.apache.sling.api.resource.*;
+import org.apache.sling.api.resource.ModifiableValueMap;
+import org.apache.sling.api.resource.PersistenceException;
+import org.apache.sling.api.resource.Resource;
+import org.apache.sling.api.resource.ResourceResolver;
+import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.api.servlets.SlingAllMethodsServlet;
 import org.apache.sling.servlets.annotations.SlingServletResourceTypes;
 import org.jetbrains.annotations.NotNull;
@@ -45,6 +49,7 @@ public class UpdateLabelServlet extends SlingAllMethodsServlet {
             Resource resource = resourceResolver.getResource(label);
             try {
                 if (resource != null) {
+                    LOG.debug("Update label on path '{}'", label);
                     updateLabel(request, resourceResolver, resource);
                 } else {
                     LOG.warn("Unable to get label '{}'", label);